add normalizedYear and numberOfCharacters to generated JSON
resolves #13
