+++ .gitignore
@@ -22,5 +22,6 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
+data.json
 src/data.json
 public/data.json
