+++ src/utils.ts
@@ -1,7 +1,6 @@
 import {Play} from './types';
-import data from './data.json';
 
-function normalizeYear (play: Play) {
+export function normalizeYear (play: Play) {
   const {premiered: p, printed, created} = play;
   const premiered: number = parseInt(p as string);
   const published = (premiered && printed)
@@ -16,22 +15,13 @@ function normalizeYear (play: Play) {
   }
 
   return year;
-}
+};
 
-function countCharacters(play: Play) {
+export function countCharacters (play: Play) {
   const {cast} = play;
   if (!cast) return undefined;
   return cast.reduce((num, item) => {
     const n = item.group ? item.group.length : 1;
     return num + n;
   }, 0)
-}
-
-const plays = data.map((p: Play) => {
-  const normalizedYear = normalizeYear(p);
-  const numberOfCharacters = countCharacters(p);
-  const authors = p.author ? [p.author] : p.authors || [];
-  return {...p, authors, normalizedYear, numberOfCharacters}
-});
-
-export default plays;
+};
